requirements.txt
asgiref==3.11.0
Django==5.2.10
djangorestframework==3.16.1
gunicorn==24.1.1
packaging==26.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
whitenoise==6.11.0
